requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.21
Pillow==9.5.0
boto3==1.29.7
stripe==7.8.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
bcrypt==4.1.2
python-socketio==5.10.0
eventlet==0.33.3
Werkzeug==2.3.7
email-validator==2.1.0
phonenumbers==8.13.27
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9